# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Efrica website  OpenApi  Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "khumbozimba4@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1"
        }
    ],
    "paths": {
...
        "/setting/getAllSettings": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Get list of settings",
                "description": "Returns list of settings",
                "operationId": "getAllSettings",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SettingResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/setting/addSetting": {
            "post": {
                "tags": [
                    "Settings"
                ],
                "summary": "Store new setting",
                "description": "Returns setting data",
                "operationId": "storeSetting",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSettingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Setting"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/setting/{id}/getSettingById": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Get setting information",
                "description": "Returns setting data",
                "operationId": "getSettingById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Setting id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Setting"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/setting/{id}/updateSettingById": {
            "put": {
                "tags": [
                    "Settings"
                ],
                "summary": "Update existing setting",
                "description": "Returns updated setting data",
                "operationId": "updateSetting",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Setting id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSettingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Setting"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/setting/{id}/deleteSettingById": {
            "delete": {
                "tags": [
                    "Settings"
                ],
                "summary": "Delete existing setting",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteSetting",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Setting id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Setting": {
                "title": "Setting",
                "description": "Setting model",
                "properties": {
                    "key": {
                        "title": "key",
                        "description": "key of the Setting model",
                        "example": ""
                    },
                    "value": {
                        "title": "value",
                        "description": "value of the Setting model",
                        "example": ""
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Setting"
                }
            },
...
            "SettingResource": {
                "title": "SettingResource",
                "description": "Setting resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Setting"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "SettingResource"
                }
            },
...
            "StoreSettingRequest": {
                "title": "Store Setting request",
                "description": "Store Setting request body data",
                "required": [
                    "key",
                    "value"
                ],
                "properties": {
                    "key": {
                        "title": "key",
                        "description": "key of the new Setting",
                        "type": "string",
                        "example": "Your example value here"
                    },
                    "value": {
                        "title": "value",
                        "description": "value of the new Setting",
                        "type": "string",
                        "example": "Your example value here"
                    }
                },
                "type": "object"
            },
...
            "UpdateSettingRequest": {
                "title": "UpdateSetting request",
                "description": "UpdateSetting request body data",
                "required": [
                    "key",
                    "value"
                ],
                "properties": {
                    "key": {
                        "title": "key",
                        "description": "key of the new Setting",
                        "type": "string",
                        "example": "Your example value here"
                    },
                    "value": {
                        "title": "value",
                        "description": "value of the new Setting",
                        "type": "string",
                        "example": "Your example value here"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}